refactor: Reorganize vite plugins
