+++ packages/@sanity/cli/src/actions/schema/vite/__tests__/plugin-schema-extraction.test.ts
@@ -4,13 +4,13 @@ import {CLITelemetryStore} from '@sanity/cli-core'
 import {SchemaValidationProblemGroup} from 'sanity'
 import {afterEach, beforeEach, describe, expect, it, vi} from 'vitest'
 
-import {createMockHttpServer, createMockWatcher} from '../../../../test/testUtils.js'
-import {SchemaExtractionError} from '../../../actions/schema/utils/SchemaExtractionError.js'
+import {createMockHttpServer, createMockWatcher} from '../../../../../test/testUtils.js'
+import {SchemaExtractionError} from '../../utils/SchemaExtractionError.js'
 import {sanitySchemaExtractionPlugin} from '../plugin-schema-extraction.js'
 
 const mockRunSchemaExtraction = vi.hoisted(() => vi.fn())
 
-vi.mock('../../../actions/schema/runSchemaExtraction.js', () => ({
+vi.mock('../../runSchemaExtraction.js', () => ({
   runSchemaExtraction: mockRunSchemaExtraction,
 }))
